requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai>=1.40.0
nltk==3.9.1
PyPDF2==3.0.1
transformers==4.36.0
torch==2.1.0